move cparam into its own thing
